+++ src/main/java/uz/tsx/repository/RegionRepository.java
@@ -11,19 +11,19 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface RegionRepository extends JpaRepository<RegionEntity,Integer> {
+public interface RegionRepository extends JpaRepository<RegionEntity,Long> {
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE tsxr.name_en=:regionName",nativeQuery = true)
     RegionEntity findByRegionName(@Param("regionName")String regionName);
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE tsxr.id=:regionId AND tsxr.status <> 'DELETED'", nativeQuery = true)
-    Optional<RegionEntity> findByRegionId(@Param("regionId") Integer regionId);
+    Optional<RegionEntity> findByRegionId(@Param("regionId") Long regionId);
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE tsxr.status <> 'DELETED'", nativeQuery = true)
     List<RegionEntity> findAllRegion();
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE (tsxr.id=:parentId OR tsxr.id=:childId) AND tsxr.status <> 'DELETED'", nativeQuery = true)
-    List<RegionEntity> getRegionIdParentAndChild(@Param("parentId") Integer parentId, @Param("childId") Integer childId);
+    List<RegionEntity> getRegionIdParentAndChild(@Param("parentId") Long parentId, @Param("childId") Long childId);
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE tsxr.parent_id IS NULL AND tsxr.status <> 'DELETED'", nativeQuery = true)
     List<RegionEntity> getRegionAllTree();
@@ -36,9 +36,9 @@ public interface RegionRepository extends JpaRepository<RegionEntity,Integer> {
             "        INNER JOIN\n" +
             "        sub_region parent ON parent.id=child.parent_id\n" +
             ")UPDATE tsx_region SET status = 'DELETED' WHERE id IN(SELECT id FROM sub_region)", nativeQuery = true)
-    void regionDelete(@Param("regionId") Integer regionId);
+    void regionDelete(@Param("regionId") Long regionId);
 
 
     @Query(value = "SELECT tsxr.* FROM tsx_region tsxr WHERE tsxr.id = :regionId AND tsxr.status <> 'DELETED'", nativeQuery = true)
-    Optional<RegionEntity> getRegionId(@Param("regionId") Integer regionId);
+    Optional<RegionEntity> getRegionId(@Param("regionId") Long regionId);
 }
