+++ src/main/java/uz/tsx/repository/UserRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     @Query(value = "SELECT tsxu.* FROM tsx_user tsxu WHERE tsxu.username = :username AND tsxu.status <> 'DELETED'", nativeQuery = true)
     Optional<UserEntity> getByUsername(@Param("username") String username);
@@ -26,7 +26,7 @@ public interface UserRepository extends JpaRepository<UserEntity, Integer> {
             "            LEFT JOIN tsx_attach tsxa ON tsxu.attach_id = tsxa.id\n" +
             "            WHERE tsxu.id = :userId AND tsxu.status <> 'DELETED'\n" +
             "            AND NOT 'ADMIN' = ANY(role_enum_list)", nativeQuery = true)
-    Optional<UserInterface> getUserInformation(@Param("userId") Integer userId);
+    Optional<UserInterface> getUserInformation(@Param("userId") Long userId);
 
     @Query(value = "SELECT tsxu.*, get_region_address(tsxu.region_id) AS address,tsxa.path, tsxa.type " +
             "FROM tsx_user tsxu " +
@@ -37,5 +37,5 @@ public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
     @Modifying
     @Query(value = "UPDATE tsx_user SET status = 'DELETED' WHERE id = :userId AND status <> 'DELETED'", nativeQuery = true)
-    void userDelete(@Param("userId") Integer userId);
+    void userDelete(@Param("userId") Long userId);
 }
